refactor: do not emit "data" event when request.body is already set
